+++ CJBCheatsMenu/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added logs to simplify troubleshooting when the menu doesn't open.
+
 ## 1.32.0
 Released 18 August 2022 for SMAPI 3.14.0 or later.